// src/app/models/activity.ts
export class IActivity {
    category: string;
    city: string;
    company: string;
    country: string;
    description: string;
    end_date: string;
    start_date: string;
    id: string;
    picture: string;
    title: string;
    user_id: string;
    location: string;
    price: string;
    address: string;
    email: string;
    phone: string;
    link: string;
    mobile: string;
    longitude: string;
    latitude: string;
    score: string;
    img_small: string;
    img_large: string;
    fb_id: string;
    start_time: string;
    end_time: string;
    interested_count: string;
    attending_count: string;
    maybe_count: string;
    timezone: string;
    cover: string;
    attending: string;
    location_id: string;
    place_id: string;
    name: string;
    owner_id: string;
    owner_name: string;
    owner_picture: string;
    slogan: string;
    capacity: string;
    space: string;
    is_wishlist: boolean;
    is_free: boolean;
}